Centralize Scene_EntityTick prototype into game.h
The 27 callers all use a u32 return and match by not re-masking r0; the
definition was declared u8 (a decomp artifact). Since the body returns only
literals 0/1, correcting the definition to the canonical u32 is byte-neutral,
and one shared u32 prototype now serves the definition and all 27 callers.
Demonstrates the conform-to-canonical method for def-vs-caller signature drift.
Clean-build verified (make tidy && make && make check).

+++ src/game/sub_08009ba0.c
@@ -52,7 +52,7 @@ extern void Timer_DecrByte(struct IwramAt35E0 *p);
 extern void Entity_SpawnFromRecord(s32 a);
 extern const u32 sEntityProcA[17];
 
-u8 Scene_EntityTick(u8 *flag)
+u32 Scene_EntityTick(u8 *flag)
 {
     Scene_DisableBg2();
     EntityPool_Reset();
